Adicionando Várias funções de uma vez

+++ ToDolistBack/src/main/java/com/doido/todolistback/domain/task/servicies/impl/TaskServiceImple.java
@@ -5,27 +5,35 @@ import com.doido.todolistback.domain.task.dtos.request.RequestTaskDto;
 import com.doido.todolistback.domain.task.entity.Task;
 import com.doido.todolistback.domain.task.servicies.TaskService;
 import com.doido.todolistback.domain.task.shared.mapper.TaskMapper;
+import com.doido.todolistback.domain.user.entity.User;
 import com.doido.todolistback.infra.repositories.TaskRepository;
+import com.doido.todolistback.infra.repositories.UserRepository;
 import com.doido.todolistback.shared.exception.CustomsExceptions.TaskNotFound;
+import com.doido.todolistback.shared.exception.CustomsExceptions.UserNotFoundException;
+import com.doido.todolistback.shared.utils.SecurityUtils;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
 @Transactional
 public class TaskServiceImple implements TaskService {
 
+    private final UserRepository userRepository;
     private final TaskRepository taskRepository;
     private final TaskMapper taskMapper;
 
 
     @Override
     public PostTaskDto addTask(PostTaskDto task){
-
         Task taskEntity = taskMapper.toTaskPost(task);
+
+        taskEntity.setUser((User) SecurityUtils.getUser());
+
         taskEntity.setCompleted(false);
 
         taskRepository.save(taskEntity);
@@ -33,32 +41,53 @@ public class TaskServiceImple implements TaskService {
     }
 
     @Override
-    public PostTaskDto updateTask(Long id, PostTaskDto task){
+    public PostTaskDto updateTask(UUID id, PostTaskDto task){
 
-        Task taskEntity = taskRepository.findById(id)
-                            .orElseThrow(()->new TaskNotFound());
+        User user = (User) userRepository
+                .findByEmail(SecurityUtils.getAuthenticationEmail());
 
-        if (task.getTitle() != null && !task.getTitle().equals("")){
-            taskEntity.setTitle(task.getTitle());
+        Task taskEntity = taskRepository
+                .findById(id)
+                .orElseThrow(()->new TaskNotFound());
+
+        if (taskEntity.getUser().equals(user)){
+            if (task.getTitle() != null && !task.getTitle().equals("")){
+                taskEntity.setTitle(task.getTitle());
+            }
+            if (task.getDescription() != null && !task.getDescription().equals("")){
+                taskEntity.setDescription(task.getDescription());
+            }
+
+            taskRepository.save(taskEntity);
+
+            return taskMapper.toPostTaskDto(taskEntity);
+        } else {
+            throw new UserNotFoundException("");
         }
-        if (task.getDescription() != null && !task.getDescription().equals("")){
-            taskEntity.setDescription(task.getDescription());
-        }
-
-        taskRepository.save(taskEntity);
-
-        return taskMapper.toPostTaskDto(taskEntity);
     }
 
     @Override
-    public void deleteTask(Long id){
-        taskRepository.findById(id).orElseThrow(() -> new TaskNotFound());
-        taskRepository.deleteById(id);
+    public void deleteTask(UUID id){
+
+        User user = (User) SecurityUtils.getUser();
+
+        Task task = taskRepository.findById(id)
+                .orElseThrow(() -> new TaskNotFound());
+
+        if (task.getUser().getId().equals(user.getId())){
+            taskRepository.deleteById(id);
+        } else {throw new UserNotFoundException("The user is not the owner of the task.");}
+
+
+
     }
 
     @Override
-    public List<RequestTaskDto> findAll(){
-        List<Task> task = taskRepository.findAll();
+    public List<RequestTaskDto> findUserTask(){
+
+        User user = (User) SecurityUtils.getUser();
+
+        List<Task> task = taskRepository.findByUserId(user.getId());
 
         List<RequestTaskDto> requestTaskDto = task.stream()
             .map(taskMapper::toRequestTaskDto)
